requirements/base.txt
# Django
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
django-filter==23.5
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.2.1

# Security
djoser==2.2.2
djangorestframework-simplejwt==5.3.1
oauthlib==3.2.2
PyJWT==2.8.0
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cryptography==42.0.2
python3-openid==3.2.0

# Redis
django-redis==5.4.0
redis==5.0.2

# Monitoring
django-prometheus==2.3.1

# Utilities
pillow==10.2.0
django-import-export==4.0.7
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.8.0rc2
inflection==0.5.1
idna==3.6
mccabe==0.7.0
packaging==23.2
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
urllib3==2.2.0
cffi==1.16.0
pytz==2023.4
tzdata==2023.4

# Parsing and Validation
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Markdown==3.5.2
sqlparse==0.4.4
swagger-spec-validator==3.0.3

# Other
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.17.1
uritemplate==4.1.1
typing_extensions==4.9.0
async-timeout==4.0.3
